feat: add bundled resources
For the purposes of the open source release, bundle resources with the app for all platforms.

This makes it easier to evaluate the project without needing to create and configure a Firebase project for storage hosting.

Note: this commit does not actually *use* bundled resources for any platform. Doing so requires setting `var remoteEndpointMode` in the `Factory*` object.